refactor: standardize seeder formatting and enhance RandomException handling
- Renamed `DatabaseSeeder` to `TestDatabaseSeeder` and added `RandomException` handling annotations.
- Replaced `rand()` with `random_int()` for better randomness and consistency.
- Aligned seeder array formatting across `DefaultUserSeeder` and `DefaultAccountCategoriesSeeder`.
- Cleaned up whitespace and improved property alignment for readability.
- Fixed `if` statement spacing and updated default values in seeders.

feat: add `AccountSeeder` and restructure `DatabaseSeeder`

- Introduced `AccountSeeder` for creating default user accounts with linked categories and properties.
- Replaced `TestDatabaseSeeder` with a more structured `DatabaseSeeder` to streamline seeding process.
- Aligned formatting in `TestDatabaseSeeder` for improved readability and consistency.
